- changed delta time to use float instead of int - removed unused shaders and cameras - replaced sun textures with shader generated sun - updated sun shader (old shader was messing with the final billboard position) - scaled everything down to regain some z buffer precision - unified camera classes - minor refactoring

+++ src/planetZoooom/interfaces/Game.java
@@ -3,5 +3,5 @@ package planetZoooom.interfaces;
 public interface Game 
 {
     void init();
-    void update(int deltaTime);
+    void update(float deltaTime);
 }
